Add student and teacher login functionality
Use bcrypt to validate passwords for both studentLogin and teacherLogin,
returning appropriate success or error messages for each authentication flow.
