--- user manegment service/controllers/loginTeacher.js
@@ -1,17 +1,0 @@
-import UserTeacher from '../models/teacherModel.js';
-
-export const loginTeacher = async (req, res) => {
-  const { firstName, password } = req.body;
-
-  try {
-    const user = await UserTeacher.findOne({ where: { firstName, password } });
-    if (!user) {
-      return res.status(400).json({ success: false, message: 'Invalid credentials' });
-    }
-    res.status(200).json({ success: true, message: `Welcome ${user.role}` });
-  } catch (error) {
-    res.status(500).json({ success: false, message: error.message });
-  }
-}
-
-export default loginTeacher;
